third_party/blink/web_tests/external/wpt/shadow-dom/declarative/getinnerhtml-expected.txt

This is a testharness.js-based test.
Found 330 tests; 186 PASS, 144 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS Element.getInnerHTML() on <a>
PASS ShadowRoot.getInnerHTML() on <a>
PASS Element.getInnerHTML() on <abbr>
PASS ShadowRoot.getInnerHTML() on <abbr>
PASS Element.getInnerHTML() on <address>
PASS ShadowRoot.getInnerHTML() on <address>
PASS Element.getInnerHTML() on <area>
PASS ShadowRoot.getInnerHTML() on <area>
FAIL Element.getInnerHTML() on <article>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<article><template shadowrootmode=\"open\"><slot></slot></template></article>" but got "<article><template shadowroot=\"open\"><slot></slot></template></article>"
FAIL Element.getInnerHTML() on <article>, with mode=closed, delegatesFocus=false. assert_equals: expected "<article><template shadowrootmode=\"closed\"><slot></slot></template></article>" but got "<article><template shadowroot=\"closed\"><slot></slot></template></article>"
FAIL Element.getInnerHTML() on <article>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<article><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></article>" but got "<article><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></article>"
FAIL Element.getInnerHTML() on <article>, with mode=closed, delegatesFocus=true. assert_equals: expected "<article><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></article>" but got "<article><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></article>"
FAIL ShadowRoot.getInnerHTML() on <article>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<article><template shadowrootmode=\"open\"><slot></slot></template></article>" but got "<article><template shadowroot=\"open\"><slot></slot></template></article>"
FAIL ShadowRoot.getInnerHTML() on <article>, with mode=closed, delegatesFocus=false. assert_equals: expected "<article><template shadowrootmode=\"closed\"><slot></slot></template></article>" but got "<article><template shadowroot=\"closed\"><slot></slot></template></article>"
FAIL ShadowRoot.getInnerHTML() on <article>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<article><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></article>" but got "<article><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></article>"
FAIL ShadowRoot.getInnerHTML() on <article>, with mode=closed, delegatesFocus=true. assert_equals: expected "<article><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></article>" but got "<article><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></article>"
FAIL Element.getInnerHTML() on <aside>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<aside><template shadowrootmode=\"open\"><slot></slot></template></aside>" but got "<aside><template shadowroot=\"open\"><slot></slot></template></aside>"
FAIL Element.getInnerHTML() on <aside>, with mode=closed, delegatesFocus=false. assert_equals: expected "<aside><template shadowrootmode=\"closed\"><slot></slot></template></aside>" but got "<aside><template shadowroot=\"closed\"><slot></slot></template></aside>"
FAIL Element.getInnerHTML() on <aside>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<aside><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></aside>" but got "<aside><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></aside>"
FAIL Element.getInnerHTML() on <aside>, with mode=closed, delegatesFocus=true. assert_equals: expected "<aside><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></aside>" but got "<aside><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></aside>"
FAIL ShadowRoot.getInnerHTML() on <aside>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<aside><template shadowrootmode=\"open\"><slot></slot></template></aside>" but got "<aside><template shadowroot=\"open\"><slot></slot></template></aside>"
FAIL ShadowRoot.getInnerHTML() on <aside>, with mode=closed, delegatesFocus=false. assert_equals: expected "<aside><template shadowrootmode=\"closed\"><slot></slot></template></aside>" but got "<aside><template shadowroot=\"closed\"><slot></slot></template></aside>"
FAIL ShadowRoot.getInnerHTML() on <aside>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<aside><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></aside>" but got "<aside><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></aside>"
FAIL ShadowRoot.getInnerHTML() on <aside>, with mode=closed, delegatesFocus=true. assert_equals: expected "<aside><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></aside>" but got "<aside><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></aside>"
PASS Element.getInnerHTML() on <audio>
PASS ShadowRoot.getInnerHTML() on <audio>
PASS Element.getInnerHTML() on <b>
PASS ShadowRoot.getInnerHTML() on <b>
PASS Element.getInnerHTML() on <base>
PASS ShadowRoot.getInnerHTML() on <base>
PASS Element.getInnerHTML() on <bdi>
PASS ShadowRoot.getInnerHTML() on <bdi>
PASS Element.getInnerHTML() on <bdo>
PASS ShadowRoot.getInnerHTML() on <bdo>
FAIL Element.getInnerHTML() on <blockquote>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<blockquote><template shadowrootmode=\"open\"><slot></slot></template></blockquote>" but got "<blockquote><template shadowroot=\"open\"><slot></slot></template></blockquote>"
FAIL Element.getInnerHTML() on <blockquote>, with mode=closed, delegatesFocus=false. assert_equals: expected "<blockquote><template shadowrootmode=\"closed\"><slot></slot></template></blockquote>" but got "<blockquote><template shadowroot=\"closed\"><slot></slot></template></blockquote>"
FAIL Element.getInnerHTML() on <blockquote>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<blockquote><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></blockquote>" but got "<blockquote><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></blockquote>"
FAIL Element.getInnerHTML() on <blockquote>, with mode=closed, delegatesFocus=true. assert_equals: expected "<blockquote><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></blockquote>" but got "<blockquote><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></blockquote>"
FAIL ShadowRoot.getInnerHTML() on <blockquote>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<blockquote><template shadowrootmode=\"open\"><slot></slot></template></blockquote>" but got "<blockquote><template shadowroot=\"open\"><slot></slot></template></blockquote>"
FAIL ShadowRoot.getInnerHTML() on <blockquote>, with mode=closed, delegatesFocus=false. assert_equals: expected "<blockquote><template shadowrootmode=\"closed\"><slot></slot></template></blockquote>" but got "<blockquote><template shadowroot=\"closed\"><slot></slot></template></blockquote>"
FAIL ShadowRoot.getInnerHTML() on <blockquote>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<blockquote><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></blockquote>" but got "<blockquote><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></blockquote>"
FAIL ShadowRoot.getInnerHTML() on <blockquote>, with mode=closed, delegatesFocus=true. assert_equals: expected "<blockquote><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></blockquote>" but got "<blockquote><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></blockquote>"
FAIL Element.getInnerHTML() on <body>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<body><template shadowrootmode=\"open\"><slot></slot></template></body>" but got "<body><template shadowroot=\"open\"><slot></slot></template></body>"
FAIL Element.getInnerHTML() on <body>, with mode=closed, delegatesFocus=false. assert_equals: expected "<body><template shadowrootmode=\"closed\"><slot></slot></template></body>" but got "<body><template shadowroot=\"closed\"><slot></slot></template></body>"
FAIL Element.getInnerHTML() on <body>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<body><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></body>" but got "<body><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></body>"
FAIL Element.getInnerHTML() on <body>, with mode=closed, delegatesFocus=true. assert_equals: expected "<body><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></body>" but got "<body><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></body>"
FAIL ShadowRoot.getInnerHTML() on <body>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<body><template shadowrootmode=\"open\"><slot></slot></template></body>" but got "<body><template shadowroot=\"open\"><slot></slot></template></body>"
FAIL ShadowRoot.getInnerHTML() on <body>, with mode=closed, delegatesFocus=false. assert_equals: expected "<body><template shadowrootmode=\"closed\"><slot></slot></template></body>" but got "<body><template shadowroot=\"closed\"><slot></slot></template></body>"
FAIL ShadowRoot.getInnerHTML() on <body>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<body><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></body>" but got "<body><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></body>"
FAIL ShadowRoot.getInnerHTML() on <body>, with mode=closed, delegatesFocus=true. assert_equals: expected "<body><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></body>" but got "<body><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></body>"
PASS Element.getInnerHTML() on <br>
PASS ShadowRoot.getInnerHTML() on <br>
PASS Element.getInnerHTML() on <button>
PASS ShadowRoot.getInnerHTML() on <button>
PASS Element.getInnerHTML() on <canvas>
PASS ShadowRoot.getInnerHTML() on <canvas>
PASS Element.getInnerHTML() on <caption>
PASS ShadowRoot.getInnerHTML() on <caption>
PASS Element.getInnerHTML() on <cite>
PASS ShadowRoot.getInnerHTML() on <cite>
PASS Element.getInnerHTML() on <code>
PASS ShadowRoot.getInnerHTML() on <code>
PASS Element.getInnerHTML() on <col>
PASS ShadowRoot.getInnerHTML() on <col>
PASS Element.getInnerHTML() on <colgroup>
PASS ShadowRoot.getInnerHTML() on <colgroup>
PASS Element.getInnerHTML() on <data>
PASS ShadowRoot.getInnerHTML() on <data>
PASS Element.getInnerHTML() on <datalist>
PASS ShadowRoot.getInnerHTML() on <datalist>
PASS Element.getInnerHTML() on <dd>
PASS ShadowRoot.getInnerHTML() on <dd>
PASS Element.getInnerHTML() on <del>
PASS ShadowRoot.getInnerHTML() on <del>
PASS Element.getInnerHTML() on <details>
PASS ShadowRoot.getInnerHTML() on <details>
PASS Element.getInnerHTML() on <dfn>
PASS ShadowRoot.getInnerHTML() on <dfn>
PASS Element.getInnerHTML() on <dialog>
PASS ShadowRoot.getInnerHTML() on <dialog>
FAIL Element.getInnerHTML() on <div>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<div><template shadowrootmode=\"open\"><slot></slot></template></div>" but got "<div><template shadowroot=\"open\"><slot></slot></template></div>"
FAIL Element.getInnerHTML() on <div>, with mode=closed, delegatesFocus=false. assert_equals: expected "<div><template shadowrootmode=\"closed\"><slot></slot></template></div>" but got "<div><template shadowroot=\"closed\"><slot></slot></template></div>"
FAIL Element.getInnerHTML() on <div>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<div><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></div>" but got "<div><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></div>"
FAIL Element.getInnerHTML() on <div>, with mode=closed, delegatesFocus=true. assert_equals: expected "<div><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></div>" but got "<div><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></div>"
FAIL ShadowRoot.getInnerHTML() on <div>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<div><template shadowrootmode=\"open\"><slot></slot></template></div>" but got "<div><template shadowroot=\"open\"><slot></slot></template></div>"
FAIL ShadowRoot.getInnerHTML() on <div>, with mode=closed, delegatesFocus=false. assert_equals: expected "<div><template shadowrootmode=\"closed\"><slot></slot></template></div>" but got "<div><template shadowroot=\"closed\"><slot></slot></template></div>"
FAIL ShadowRoot.getInnerHTML() on <div>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<div><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></div>" but got "<div><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></div>"
FAIL ShadowRoot.getInnerHTML() on <div>, with mode=closed, delegatesFocus=true. assert_equals: expected "<div><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></div>" but got "<div><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></div>"
PASS Element.getInnerHTML() on <dl>
PASS ShadowRoot.getInnerHTML() on <dl>
PASS Element.getInnerHTML() on <dt>
PASS ShadowRoot.getInnerHTML() on <dt>
PASS Element.getInnerHTML() on <em>
PASS ShadowRoot.getInnerHTML() on <em>
PASS Element.getInnerHTML() on <embed>
PASS ShadowRoot.getInnerHTML() on <embed>
PASS Element.getInnerHTML() on <fieldset>
PASS ShadowRoot.getInnerHTML() on <fieldset>
PASS Element.getInnerHTML() on <figcaption>
PASS ShadowRoot.getInnerHTML() on <figcaption>
PASS Element.getInnerHTML() on <figure>
PASS ShadowRoot.getInnerHTML() on <figure>
FAIL Element.getInnerHTML() on <footer>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<footer><template shadowrootmode=\"open\"><slot></slot></template></footer>" but got "<footer><template shadowroot=\"open\"><slot></slot></template></footer>"
FAIL Element.getInnerHTML() on <footer>, with mode=closed, delegatesFocus=false. assert_equals: expected "<footer><template shadowrootmode=\"closed\"><slot></slot></template></footer>" but got "<footer><template shadowroot=\"closed\"><slot></slot></template></footer>"
FAIL Element.getInnerHTML() on <footer>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<footer><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></footer>" but got "<footer><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></footer>"
FAIL Element.getInnerHTML() on <footer>, with mode=closed, delegatesFocus=true. assert_equals: expected "<footer><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></footer>" but got "<footer><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></footer>"
FAIL ShadowRoot.getInnerHTML() on <footer>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<footer><template shadowrootmode=\"open\"><slot></slot></template></footer>" but got "<footer><template shadowroot=\"open\"><slot></slot></template></footer>"
FAIL ShadowRoot.getInnerHTML() on <footer>, with mode=closed, delegatesFocus=false. assert_equals: expected "<footer><template shadowrootmode=\"closed\"><slot></slot></template></footer>" but got "<footer><template shadowroot=\"closed\"><slot></slot></template></footer>"
FAIL ShadowRoot.getInnerHTML() on <footer>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<footer><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></footer>" but got "<footer><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></footer>"
FAIL ShadowRoot.getInnerHTML() on <footer>, with mode=closed, delegatesFocus=true. assert_equals: expected "<footer><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></footer>" but got "<footer><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></footer>"
PASS Element.getInnerHTML() on <form>
PASS ShadowRoot.getInnerHTML() on <form>
FAIL Element.getInnerHTML() on <h1>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<h1><template shadowrootmode=\"open\"><slot></slot></template></h1>" but got "<h1><template shadowroot=\"open\"><slot></slot></template></h1>"
FAIL Element.getInnerHTML() on <h1>, with mode=closed, delegatesFocus=false. assert_equals: expected "<h1><template shadowrootmode=\"closed\"><slot></slot></template></h1>" but got "<h1><template shadowroot=\"closed\"><slot></slot></template></h1>"
FAIL Element.getInnerHTML() on <h1>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<h1><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h1>" but got "<h1><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h1>"
FAIL Element.getInnerHTML() on <h1>, with mode=closed, delegatesFocus=true. assert_equals: expected "<h1><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h1>" but got "<h1><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h1>"
FAIL ShadowRoot.getInnerHTML() on <h1>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<h1><template shadowrootmode=\"open\"><slot></slot></template></h1>" but got "<h1><template shadowroot=\"open\"><slot></slot></template></h1>"
FAIL ShadowRoot.getInnerHTML() on <h1>, with mode=closed, delegatesFocus=false. assert_equals: expected "<h1><template shadowrootmode=\"closed\"><slot></slot></template></h1>" but got "<h1><template shadowroot=\"closed\"><slot></slot></template></h1>"
FAIL ShadowRoot.getInnerHTML() on <h1>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<h1><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h1>" but got "<h1><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h1>"
FAIL ShadowRoot.getInnerHTML() on <h1>, with mode=closed, delegatesFocus=true. assert_equals: expected "<h1><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h1>" but got "<h1><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h1>"
FAIL Element.getInnerHTML() on <h2>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<h2><template shadowrootmode=\"open\"><slot></slot></template></h2>" but got "<h2><template shadowroot=\"open\"><slot></slot></template></h2>"
FAIL Element.getInnerHTML() on <h2>, with mode=closed, delegatesFocus=false. assert_equals: expected "<h2><template shadowrootmode=\"closed\"><slot></slot></template></h2>" but got "<h2><template shadowroot=\"closed\"><slot></slot></template></h2>"
FAIL Element.getInnerHTML() on <h2>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<h2><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h2>" but got "<h2><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h2>"
FAIL Element.getInnerHTML() on <h2>, with mode=closed, delegatesFocus=true. assert_equals: expected "<h2><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h2>" but got "<h2><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h2>"
FAIL ShadowRoot.getInnerHTML() on <h2>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<h2><template shadowrootmode=\"open\"><slot></slot></template></h2>" but got "<h2><template shadowroot=\"open\"><slot></slot></template></h2>"
FAIL ShadowRoot.getInnerHTML() on <h2>, with mode=closed, delegatesFocus=false. assert_equals: expected "<h2><template shadowrootmode=\"closed\"><slot></slot></template></h2>" but got "<h2><template shadowroot=\"closed\"><slot></slot></template></h2>"
FAIL ShadowRoot.getInnerHTML() on <h2>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<h2><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h2>" but got "<h2><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h2>"
FAIL ShadowRoot.getInnerHTML() on <h2>, with mode=closed, delegatesFocus=true. assert_equals: expected "<h2><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h2>" but got "<h2><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h2>"
FAIL Element.getInnerHTML() on <h3>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<h3><template shadowrootmode=\"open\"><slot></slot></template></h3>" but got "<h3><template shadowroot=\"open\"><slot></slot></template></h3>"
FAIL Element.getInnerHTML() on <h3>, with mode=closed, delegatesFocus=false. assert_equals: expected "<h3><template shadowrootmode=\"closed\"><slot></slot></template></h3>" but got "<h3><template shadowroot=\"closed\"><slot></slot></template></h3>"
FAIL Element.getInnerHTML() on <h3>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<h3><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h3>" but got "<h3><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h3>"
FAIL Element.getInnerHTML() on <h3>, with mode=closed, delegatesFocus=true. assert_equals: expected "<h3><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h3>" but got "<h3><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h3>"
FAIL ShadowRoot.getInnerHTML() on <h3>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<h3><template shadowrootmode=\"open\"><slot></slot></template></h3>" but got "<h3><template shadowroot=\"open\"><slot></slot></template></h3>"
FAIL ShadowRoot.getInnerHTML() on <h3>, with mode=closed, delegatesFocus=false. assert_equals: expected "<h3><template shadowrootmode=\"closed\"><slot></slot></template></h3>" but got "<h3><template shadowroot=\"closed\"><slot></slot></template></h3>"
FAIL ShadowRoot.getInnerHTML() on <h3>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<h3><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h3>" but got "<h3><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h3>"
FAIL ShadowRoot.getInnerHTML() on <h3>, with mode=closed, delegatesFocus=true. assert_equals: expected "<h3><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h3>" but got "<h3><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h3>"
FAIL Element.getInnerHTML() on <h4>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<h4><template shadowrootmode=\"open\"><slot></slot></template></h4>" but got "<h4><template shadowroot=\"open\"><slot></slot></template></h4>"
FAIL Element.getInnerHTML() on <h4>, with mode=closed, delegatesFocus=false. assert_equals: expected "<h4><template shadowrootmode=\"closed\"><slot></slot></template></h4>" but got "<h4><template shadowroot=\"closed\"><slot></slot></template></h4>"
FAIL Element.getInnerHTML() on <h4>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<h4><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h4>" but got "<h4><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h4>"
FAIL Element.getInnerHTML() on <h4>, with mode=closed, delegatesFocus=true. assert_equals: expected "<h4><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h4>" but got "<h4><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h4>"
FAIL ShadowRoot.getInnerHTML() on <h4>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<h4><template shadowrootmode=\"open\"><slot></slot></template></h4>" but got "<h4><template shadowroot=\"open\"><slot></slot></template></h4>"
FAIL ShadowRoot.getInnerHTML() on <h4>, with mode=closed, delegatesFocus=false. assert_equals: expected "<h4><template shadowrootmode=\"closed\"><slot></slot></template></h4>" but got "<h4><template shadowroot=\"closed\"><slot></slot></template></h4>"
FAIL ShadowRoot.getInnerHTML() on <h4>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<h4><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h4>" but got "<h4><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h4>"
FAIL ShadowRoot.getInnerHTML() on <h4>, with mode=closed, delegatesFocus=true. assert_equals: expected "<h4><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h4>" but got "<h4><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h4>"
FAIL Element.getInnerHTML() on <h5>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<h5><template shadowrootmode=\"open\"><slot></slot></template></h5>" but got "<h5><template shadowroot=\"open\"><slot></slot></template></h5>"
FAIL Element.getInnerHTML() on <h5>, with mode=closed, delegatesFocus=false. assert_equals: expected "<h5><template shadowrootmode=\"closed\"><slot></slot></template></h5>" but got "<h5><template shadowroot=\"closed\"><slot></slot></template></h5>"
FAIL Element.getInnerHTML() on <h5>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<h5><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h5>" but got "<h5><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h5>"
FAIL Element.getInnerHTML() on <h5>, with mode=closed, delegatesFocus=true. assert_equals: expected "<h5><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h5>" but got "<h5><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h5>"
FAIL ShadowRoot.getInnerHTML() on <h5>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<h5><template shadowrootmode=\"open\"><slot></slot></template></h5>" but got "<h5><template shadowroot=\"open\"><slot></slot></template></h5>"
FAIL ShadowRoot.getInnerHTML() on <h5>, with mode=closed, delegatesFocus=false. assert_equals: expected "<h5><template shadowrootmode=\"closed\"><slot></slot></template></h5>" but got "<h5><template shadowroot=\"closed\"><slot></slot></template></h5>"
FAIL ShadowRoot.getInnerHTML() on <h5>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<h5><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h5>" but got "<h5><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h5>"
FAIL ShadowRoot.getInnerHTML() on <h5>, with mode=closed, delegatesFocus=true. assert_equals: expected "<h5><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h5>" but got "<h5><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h5>"
FAIL Element.getInnerHTML() on <h6>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<h6><template shadowrootmode=\"open\"><slot></slot></template></h6>" but got "<h6><template shadowroot=\"open\"><slot></slot></template></h6>"
FAIL Element.getInnerHTML() on <h6>, with mode=closed, delegatesFocus=false. assert_equals: expected "<h6><template shadowrootmode=\"closed\"><slot></slot></template></h6>" but got "<h6><template shadowroot=\"closed\"><slot></slot></template></h6>"
FAIL Element.getInnerHTML() on <h6>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<h6><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h6>" but got "<h6><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h6>"
FAIL Element.getInnerHTML() on <h6>, with mode=closed, delegatesFocus=true. assert_equals: expected "<h6><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h6>" but got "<h6><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h6>"
FAIL ShadowRoot.getInnerHTML() on <h6>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<h6><template shadowrootmode=\"open\"><slot></slot></template></h6>" but got "<h6><template shadowroot=\"open\"><slot></slot></template></h6>"
FAIL ShadowRoot.getInnerHTML() on <h6>, with mode=closed, delegatesFocus=false. assert_equals: expected "<h6><template shadowrootmode=\"closed\"><slot></slot></template></h6>" but got "<h6><template shadowroot=\"closed\"><slot></slot></template></h6>"
FAIL ShadowRoot.getInnerHTML() on <h6>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<h6><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h6>" but got "<h6><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h6>"
FAIL ShadowRoot.getInnerHTML() on <h6>, with mode=closed, delegatesFocus=true. assert_equals: expected "<h6><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h6>" but got "<h6><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></h6>"
PASS Element.getInnerHTML() on <head>
PASS ShadowRoot.getInnerHTML() on <head>
FAIL Element.getInnerHTML() on <header>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<header><template shadowrootmode=\"open\"><slot></slot></template></header>" but got "<header><template shadowroot=\"open\"><slot></slot></template></header>"
FAIL Element.getInnerHTML() on <header>, with mode=closed, delegatesFocus=false. assert_equals: expected "<header><template shadowrootmode=\"closed\"><slot></slot></template></header>" but got "<header><template shadowroot=\"closed\"><slot></slot></template></header>"
FAIL Element.getInnerHTML() on <header>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<header><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></header>" but got "<header><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></header>"
FAIL Element.getInnerHTML() on <header>, with mode=closed, delegatesFocus=true. assert_equals: expected "<header><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></header>" but got "<header><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></header>"
FAIL ShadowRoot.getInnerHTML() on <header>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<header><template shadowrootmode=\"open\"><slot></slot></template></header>" but got "<header><template shadowroot=\"open\"><slot></slot></template></header>"
FAIL ShadowRoot.getInnerHTML() on <header>, with mode=closed, delegatesFocus=false. assert_equals: expected "<header><template shadowrootmode=\"closed\"><slot></slot></template></header>" but got "<header><template shadowroot=\"closed\"><slot></slot></template></header>"
FAIL ShadowRoot.getInnerHTML() on <header>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<header><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></header>" but got "<header><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></header>"
FAIL ShadowRoot.getInnerHTML() on <header>, with mode=closed, delegatesFocus=true. assert_equals: expected "<header><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></header>" but got "<header><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></header>"
PASS Element.getInnerHTML() on <hr>
PASS ShadowRoot.getInnerHTML() on <hr>
PASS Element.getInnerHTML() on <html>
PASS ShadowRoot.getInnerHTML() on <html>
PASS Element.getInnerHTML() on <i>
PASS ShadowRoot.getInnerHTML() on <i>
PASS Element.getInnerHTML() on <iframe>
PASS ShadowRoot.getInnerHTML() on <iframe>
PASS Element.getInnerHTML() on <img>
PASS ShadowRoot.getInnerHTML() on <img>
PASS Element.getInnerHTML() on <input>
PASS ShadowRoot.getInnerHTML() on <input>
PASS Element.getInnerHTML() on <ins>
PASS ShadowRoot.getInnerHTML() on <ins>
PASS Element.getInnerHTML() on <kbd>
PASS ShadowRoot.getInnerHTML() on <kbd>
PASS Element.getInnerHTML() on <label>
PASS ShadowRoot.getInnerHTML() on <label>
PASS Element.getInnerHTML() on <legend>
PASS ShadowRoot.getInnerHTML() on <legend>
PASS Element.getInnerHTML() on <li>
PASS ShadowRoot.getInnerHTML() on <li>
PASS Element.getInnerHTML() on <link>
PASS ShadowRoot.getInnerHTML() on <link>
FAIL Element.getInnerHTML() on <main>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<main><template shadowrootmode=\"open\"><slot></slot></template></main>" but got "<main><template shadowroot=\"open\"><slot></slot></template></main>"
FAIL Element.getInnerHTML() on <main>, with mode=closed, delegatesFocus=false. assert_equals: expected "<main><template shadowrootmode=\"closed\"><slot></slot></template></main>" but got "<main><template shadowroot=\"closed\"><slot></slot></template></main>"
FAIL Element.getInnerHTML() on <main>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<main><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></main>" but got "<main><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></main>"
FAIL Element.getInnerHTML() on <main>, with mode=closed, delegatesFocus=true. assert_equals: expected "<main><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></main>" but got "<main><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></main>"
FAIL ShadowRoot.getInnerHTML() on <main>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<main><template shadowrootmode=\"open\"><slot></slot></template></main>" but got "<main><template shadowroot=\"open\"><slot></slot></template></main>"
FAIL ShadowRoot.getInnerHTML() on <main>, with mode=closed, delegatesFocus=false. assert_equals: expected "<main><template shadowrootmode=\"closed\"><slot></slot></template></main>" but got "<main><template shadowroot=\"closed\"><slot></slot></template></main>"
FAIL ShadowRoot.getInnerHTML() on <main>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<main><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></main>" but got "<main><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></main>"
FAIL ShadowRoot.getInnerHTML() on <main>, with mode=closed, delegatesFocus=true. assert_equals: expected "<main><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></main>" but got "<main><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></main>"
PASS Element.getInnerHTML() on <map>
PASS ShadowRoot.getInnerHTML() on <map>
PASS Element.getInnerHTML() on <mark>
PASS ShadowRoot.getInnerHTML() on <mark>
PASS Element.getInnerHTML() on <menu>
PASS ShadowRoot.getInnerHTML() on <menu>
PASS Element.getInnerHTML() on <meta>
PASS ShadowRoot.getInnerHTML() on <meta>
PASS Element.getInnerHTML() on <meter>
PASS ShadowRoot.getInnerHTML() on <meter>
FAIL Element.getInnerHTML() on <nav>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<nav><template shadowrootmode=\"open\"><slot></slot></template></nav>" but got "<nav><template shadowroot=\"open\"><slot></slot></template></nav>"
FAIL Element.getInnerHTML() on <nav>, with mode=closed, delegatesFocus=false. assert_equals: expected "<nav><template shadowrootmode=\"closed\"><slot></slot></template></nav>" but got "<nav><template shadowroot=\"closed\"><slot></slot></template></nav>"
FAIL Element.getInnerHTML() on <nav>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<nav><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></nav>" but got "<nav><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></nav>"
FAIL Element.getInnerHTML() on <nav>, with mode=closed, delegatesFocus=true. assert_equals: expected "<nav><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></nav>" but got "<nav><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></nav>"
FAIL ShadowRoot.getInnerHTML() on <nav>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<nav><template shadowrootmode=\"open\"><slot></slot></template></nav>" but got "<nav><template shadowroot=\"open\"><slot></slot></template></nav>"
FAIL ShadowRoot.getInnerHTML() on <nav>, with mode=closed, delegatesFocus=false. assert_equals: expected "<nav><template shadowrootmode=\"closed\"><slot></slot></template></nav>" but got "<nav><template shadowroot=\"closed\"><slot></slot></template></nav>"
FAIL ShadowRoot.getInnerHTML() on <nav>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<nav><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></nav>" but got "<nav><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></nav>"
FAIL ShadowRoot.getInnerHTML() on <nav>, with mode=closed, delegatesFocus=true. assert_equals: expected "<nav><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></nav>" but got "<nav><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></nav>"
PASS Element.getInnerHTML() on <noscript>
PASS ShadowRoot.getInnerHTML() on <noscript>
PASS Element.getInnerHTML() on <object>
PASS ShadowRoot.getInnerHTML() on <object>
PASS Element.getInnerHTML() on <ol>
PASS ShadowRoot.getInnerHTML() on <ol>
PASS Element.getInnerHTML() on <optgroup>
PASS ShadowRoot.getInnerHTML() on <optgroup>
PASS Element.getInnerHTML() on <option>
PASS ShadowRoot.getInnerHTML() on <option>
PASS Element.getInnerHTML() on <output>
PASS ShadowRoot.getInnerHTML() on <output>
FAIL Element.getInnerHTML() on <p>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<p><template shadowrootmode=\"open\"><slot></slot></template></p>" but got "<p><template shadowroot=\"open\"><slot></slot></template></p>"
FAIL Element.getInnerHTML() on <p>, with mode=closed, delegatesFocus=false. assert_equals: expected "<p><template shadowrootmode=\"closed\"><slot></slot></template></p>" but got "<p><template shadowroot=\"closed\"><slot></slot></template></p>"
FAIL Element.getInnerHTML() on <p>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<p><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></p>" but got "<p><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></p>"
FAIL Element.getInnerHTML() on <p>, with mode=closed, delegatesFocus=true. assert_equals: expected "<p><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></p>" but got "<p><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></p>"
FAIL ShadowRoot.getInnerHTML() on <p>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<p><template shadowrootmode=\"open\"><slot></slot></template></p>" but got "<p><template shadowroot=\"open\"><slot></slot></template></p>"
FAIL ShadowRoot.getInnerHTML() on <p>, with mode=closed, delegatesFocus=false. assert_equals: expected "<p><template shadowrootmode=\"closed\"><slot></slot></template></p>" but got "<p><template shadowroot=\"closed\"><slot></slot></template></p>"
FAIL ShadowRoot.getInnerHTML() on <p>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<p><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></p>" but got "<p><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></p>"
FAIL ShadowRoot.getInnerHTML() on <p>, with mode=closed, delegatesFocus=true. assert_equals: expected "<p><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></p>" but got "<p><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></p>"
PASS Element.getInnerHTML() on <param>
PASS ShadowRoot.getInnerHTML() on <param>
PASS Element.getInnerHTML() on <pre>
PASS ShadowRoot.getInnerHTML() on <pre>
PASS Element.getInnerHTML() on <progress>
PASS ShadowRoot.getInnerHTML() on <progress>
PASS Element.getInnerHTML() on <q>
PASS ShadowRoot.getInnerHTML() on <q>
PASS Element.getInnerHTML() on <rp>
PASS ShadowRoot.getInnerHTML() on <rp>
PASS Element.getInnerHTML() on <rt>
PASS ShadowRoot.getInnerHTML() on <rt>
PASS Element.getInnerHTML() on <ruby>
PASS ShadowRoot.getInnerHTML() on <ruby>
PASS Element.getInnerHTML() on <s>
PASS ShadowRoot.getInnerHTML() on <s>
PASS Element.getInnerHTML() on <samp>
PASS ShadowRoot.getInnerHTML() on <samp>
PASS Element.getInnerHTML() on <script>
PASS ShadowRoot.getInnerHTML() on <script>
FAIL Element.getInnerHTML() on <section>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<section><template shadowrootmode=\"open\"><slot></slot></template></section>" but got "<section><template shadowroot=\"open\"><slot></slot></template></section>"
FAIL Element.getInnerHTML() on <section>, with mode=closed, delegatesFocus=false. assert_equals: expected "<section><template shadowrootmode=\"closed\"><slot></slot></template></section>" but got "<section><template shadowroot=\"closed\"><slot></slot></template></section>"
FAIL Element.getInnerHTML() on <section>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<section><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></section>" but got "<section><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></section>"
FAIL Element.getInnerHTML() on <section>, with mode=closed, delegatesFocus=true. assert_equals: expected "<section><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></section>" but got "<section><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></section>"
FAIL ShadowRoot.getInnerHTML() on <section>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<section><template shadowrootmode=\"open\"><slot></slot></template></section>" but got "<section><template shadowroot=\"open\"><slot></slot></template></section>"
FAIL ShadowRoot.getInnerHTML() on <section>, with mode=closed, delegatesFocus=false. assert_equals: expected "<section><template shadowrootmode=\"closed\"><slot></slot></template></section>" but got "<section><template shadowroot=\"closed\"><slot></slot></template></section>"
FAIL ShadowRoot.getInnerHTML() on <section>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<section><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></section>" but got "<section><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></section>"
FAIL ShadowRoot.getInnerHTML() on <section>, with mode=closed, delegatesFocus=true. assert_equals: expected "<section><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></section>" but got "<section><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></section>"
PASS Element.getInnerHTML() on <select>
PASS ShadowRoot.getInnerHTML() on <select>
PASS Element.getInnerHTML() on <slot>
PASS ShadowRoot.getInnerHTML() on <slot>
PASS Element.getInnerHTML() on <small>
PASS ShadowRoot.getInnerHTML() on <small>
PASS Element.getInnerHTML() on <source>
PASS ShadowRoot.getInnerHTML() on <source>
FAIL Element.getInnerHTML() on <span>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<span><template shadowrootmode=\"open\"><slot></slot></template></span>" but got "<span><template shadowroot=\"open\"><slot></slot></template></span>"
FAIL Element.getInnerHTML() on <span>, with mode=closed, delegatesFocus=false. assert_equals: expected "<span><template shadowrootmode=\"closed\"><slot></slot></template></span>" but got "<span><template shadowroot=\"closed\"><slot></slot></template></span>"
FAIL Element.getInnerHTML() on <span>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<span><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></span>" but got "<span><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></span>"
FAIL Element.getInnerHTML() on <span>, with mode=closed, delegatesFocus=true. assert_equals: expected "<span><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></span>" but got "<span><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></span>"
FAIL ShadowRoot.getInnerHTML() on <span>, with mode=open, delegatesFocus=false. assert_equals: The default for includeShadowRoots should be true expected "<span><template shadowrootmode=\"open\"><slot></slot></template></span>" but got "<span><template shadowroot=\"open\"><slot></slot></template></span>"
FAIL ShadowRoot.getInnerHTML() on <span>, with mode=closed, delegatesFocus=false. assert_equals: expected "<span><template shadowrootmode=\"closed\"><slot></slot></template></span>" but got "<span><template shadowroot=\"closed\"><slot></slot></template></span>"
FAIL ShadowRoot.getInnerHTML() on <span>, with mode=open, delegatesFocus=true. assert_equals: The default for includeShadowRoots should be true expected "<span><template shadowrootmode=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></span>" but got "<span><template shadowroot=\"open\" shadowrootdelegatesfocus=\"\"><slot></slot></template></span>"
FAIL ShadowRoot.getInnerHTML() on <span>, with mode=closed, delegatesFocus=true. assert_equals: expected "<span><template shadowrootmode=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></span>" but got "<span><template shadowroot=\"closed\" shadowrootdelegatesfocus=\"\"><slot></slot></template></span>"
PASS Element.getInnerHTML() on <strong>
PASS ShadowRoot.getInnerHTML() on <strong>
PASS Element.getInnerHTML() on <style>
PASS ShadowRoot.getInnerHTML() on <style>
PASS Element.getInnerHTML() on <sub>
PASS ShadowRoot.getInnerHTML() on <sub>
PASS Element.getInnerHTML() on <sup>
PASS ShadowRoot.getInnerHTML() on <sup>
PASS Element.getInnerHTML() on <summary>
PASS ShadowRoot.getInnerHTML() on <summary>
PASS Element.getInnerHTML() on <table>
PASS ShadowRoot.getInnerHTML() on <table>
PASS Element.getInnerHTML() on <tbody>
PASS ShadowRoot.getInnerHTML() on <tbody>
PASS Element.getInnerHTML() on <td>
PASS ShadowRoot.getInnerHTML() on <td>
PASS Element.getInnerHTML() on <template>
PASS ShadowRoot.getInnerHTML() on <template>
PASS Element.getInnerHTML() on <textarea>
PASS ShadowRoot.getInnerHTML() on <textarea>
PASS Element.getInnerHTML() on <tfoot>
PASS ShadowRoot.getInnerHTML() on <tfoot>
PASS Element.getInnerHTML() on <th>
PASS ShadowRoot.getInnerHTML() on <th>
PASS Element.getInnerHTML() on <thead>
PASS ShadowRoot.getInnerHTML() on <thead>
PASS Element.getInnerHTML() on <time>
PASS ShadowRoot.getInnerHTML() on <time>
PASS Element.getInnerHTML() on <title>
PASS ShadowRoot.getInnerHTML() on <title>
PASS Element.getInnerHTML() on <tr>
PASS ShadowRoot.getInnerHTML() on <tr>
PASS Element.getInnerHTML() on <track>
PASS ShadowRoot.getInnerHTML() on <track>
PASS Element.getInnerHTML() on <u>
PASS ShadowRoot.getInnerHTML() on <u>
PASS Element.getInnerHTML() on <ul>
PASS ShadowRoot.getInnerHTML() on <ul>
PASS Element.getInnerHTML() on <var>
PASS ShadowRoot.getInnerHTML() on <var>
PASS Element.getInnerHTML() on <video>
PASS ShadowRoot.getInnerHTML() on <video>
PASS Element.getInnerHTML() on <wbr>
PASS ShadowRoot.getInnerHTML() on <wbr>
PASS Element.getInnerHTML() on <htmlunknown>
PASS ShadowRoot.getInnerHTML() on <htmlunknown>
Harness: the test ran to completion.
